Replaced "long long int" with "int64_t", and "unsigned long long int" with "uint64_t" to improve portability.
git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15483 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/AST/symbol.cpp
@@ -1,5 +1,6 @@
 #define __STDC_FORMAT_MACROS
 #include <inttypes.h>
+#include <stdint.h>
 #include "astutil.h"
 #include "bb.h"
 #include "build.h"
@@ -886,7 +887,7 @@ VarSymbol* new_BoolSymbol(bool b, IF1_bool_type size) {
   return s;
 }
 
-VarSymbol *new_IntSymbol(long long int b, IF1_int_type size) {
+VarSymbol *new_IntSymbol(int64_t b, IF1_int_type size) {
   Immediate imm;
   switch (size) {
   case INT_SIZE_8  : imm.v_int8   = b; break;
@@ -910,7 +911,7 @@ VarSymbol *new_IntSymbol(long long int b, IF1_int_type size) {
   return s;
 }
 
-VarSymbol *new_UIntSymbol(unsigned long long int b, IF1_int_type size) {
+VarSymbol *new_UIntSymbol(uint64_t b, IF1_int_type size) {
   Immediate imm;
   switch (size) {
   case INT_SIZE_8  : imm.v_uint8   = b; break;
