+++ compiler/include/symbol.h
@@ -2,6 +2,7 @@
 #define _SYMBOL_H_
 
 #include <bitset>
+#include <stdint.h>
 #include "alist.h"
 #include "baseAST.h"
 #include "bitVec.h"
@@ -229,8 +230,8 @@ class LabelSymbol : public Symbol {
 
 VarSymbol *new_StringSymbol(const char *s);
 VarSymbol *new_BoolSymbol(bool b, IF1_bool_type size=BOOL_SIZE_SYS);
-VarSymbol *new_IntSymbol(long long int b, IF1_int_type size=INT_SIZE_32);
-VarSymbol *new_UIntSymbol(unsigned long long int b, IF1_int_type size=INT_SIZE_32);
+VarSymbol *new_IntSymbol(int64_t b, IF1_int_type size=INT_SIZE_32);
+VarSymbol *new_UIntSymbol(uint64_t b, IF1_int_type size=INT_SIZE_32);
 VarSymbol *new_RealSymbol(const char *n, long double b, IF1_float_type size=FLOAT_SIZE_64);
 VarSymbol *new_ImagSymbol(const char *n, long double b, IF1_float_type size=FLOAT_SIZE_64);
 VarSymbol *new_ComplexSymbol(const char *n, long double r, long double i, IF1_complex_type size=COMPLEX_SIZE_128);
